requirements.txt
pyyaml
pygit2
